Clean up performance tests
