+++ apps/performance-tests/src/hierarchies/FlatHierarchy.test.ts
@@ -3,6 +3,6 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import { runHierarchyTest } from "../util/TestUtilities";
+import { runHierarchyTest } from "./RunHierarchyTest";
 
 runHierarchyTest({ testName: "flat 50k elements list", iModelName: "50k flat elements" });
